Move all the library source codes to lib/ directory
