+++ lib/dsyevc3.c
