+++ lib/dsyevq3.c
