+++ lib/dsyevq3.h
