+++ makefile
@@ -8,7 +8,7 @@ LD_LIBS = -lm
 # CFLAGS += -fsanitize=address -g  # useful for segmentation fault
 
 # Parse all the files
-SRC = $(wildcard src/*.c)
+SRC = $(wildcard src/*.c) $(wildcard lib/*.c)
 OBJ = $(patsubst %.c,build/%.o,$(notdir $(SRC)))
 DEP = $(OBJ:%.o=%.d)
 DIR = bin/ build/
@@ -27,6 +27,9 @@ bin/treeeeee: $(OBJ)
 
 build/%.o: src/%.c
 	$(CC) $(CFLAGS) -c -MMD -o $@ $<
+
+build/%.o: lib/%.c
+	$(CC) -O2 -c -MMD -o $@ $<
 
 clean:
 	rm build/*
